refactor : 테스트코드 리팩터링 / Rank에 matchCount 게터 추가
